editar usuario

+++ Financeiro/src/controle/UsuarioBean.java
@@ -7,7 +7,6 @@ import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ViewScoped;
 import javax.faces.context.FacesContext;
 import javax.servlet.http.HttpSession;
-
 import dao.UsuarioDAO;
 import dao.UsuarioJPADAO;
 import modelo.Usuario;
@@ -18,14 +17,20 @@ public class UsuarioBean extends AbstractBean implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 	private Usuario usuario;
+	private Usuario usuarioAtualiza;
 	private List<Usuario> usuarios;
 	private String senhaAtual;
 	private String novaSenha;
 	private String confirmacaoSenha;
+	private String loginUsuario;
+	private boolean isAdmin;
+	private boolean isNaoAdmin;
 
 	public UsuarioBean() {
 		this.usuario = new Usuario();
 		this.usuarios = new ArrayList<Usuario>();
+		buscarUsuarioLogado();
+		selecionarParaAtualizar();
 	}
 
 	public void buscarTodos() {
@@ -34,37 +39,54 @@ public class UsuarioBean extends AbstractBean implements Serializable {
 		this.usuarios = usuarioDAO.find();
 	}
 
+	public void buscarUsuarioLogado() {
+		HttpSession sessao = (HttpSession) FacesContext.getCurrentInstance().getExternalContext().getSession(false);
+		String login = (String) sessao.getAttribute("login");
+		this.loginUsuario = login;
+		if (login.equals("admin")) {
+			this.isAdmin = true;
+			this.isNaoAdmin = false;
+		} else {
+			this.isAdmin = false;
+			this.isNaoAdmin = true;
+		}
+	}
+
 	public void cadastrar() {
-		try {
+		if (isAdmin) {
 			UsuarioDAO usuarioDAO = new UsuarioJPADAO();
 			usuarioDAO.save(this.usuario);
 			displayInfoMessageToUser("Cadastrado com sucesso!");
 			this.usuario = new Usuario();
-		} catch (Exception e) {
-			displayErrorMessageToUser("Erro no cadastramento!");
-			this.usuario = new Usuario();
+		} else {
+			displayErrorMessageToUser("Usuário não autorizado para fazer cadastro de usuário!");
 		}
+
 	}
 
 	public void atualizar() {
 		UsuarioDAO usuarioDAO = new UsuarioJPADAO();
-		usuarioDAO.save(this.usuario);
+		usuarioDAO.save(this.usuarioAtualiza);
 		displayInfoMessageToUser("Atualizado com sucesso!");
-		this.usuario = new Usuario();
+		this.usuarioAtualiza = new Usuario();
 	}
 
 	public void excluir(Usuario u) {
-		UsuarioDAO usuarioDAO = new UsuarioJPADAO();
-		usuarioDAO.delete(u);
-		displayErrorMessageToUser("Excluido com sucesso!");
-		this.usuarios.remove(u);
+		if (isAdmin) {
+			UsuarioDAO usuarioDAO = new UsuarioJPADAO();
+			usuarioDAO.delete(u);
+			displayErrorMessageToUser("Excluido com sucesso!");
+			this.usuarios.remove(u);
+		} else {
+			displayErrorMessageToUser("Usuário não autorizado para fazer exclusão de usuário!");
+		}
 	}
 
-	public void selecionarPerfil() {
+	public void selecionarParaAtualizar() {
 		HttpSession sessao = (HttpSession) FacesContext.getCurrentInstance().getExternalContext().getSession(false);
 		String login = (String) sessao.getAttribute("login");
 		UsuarioDAO uDAO = new UsuarioJPADAO();
-		this.usuario = uDAO.find(login);
+		this.usuarioAtualiza = uDAO.find(login);
 	}
 
 	public void alterarSenha() {
@@ -140,4 +162,37 @@ public class UsuarioBean extends AbstractBean implements Serializable {
 	public void setConfirmacaoSenha(String confirmacaoSenha) {
 		this.confirmacaoSenha = confirmacaoSenha;
 	}
+
+	public String getLoginUsuario() {
+		return loginUsuario;
+	}
+
+	public void setLoginUsuario(String loginUsuario) {
+		this.loginUsuario = loginUsuario;
+	}
+
+	public boolean isNaoAdmin() {
+		return isNaoAdmin;
+	}
+
+	public void setNaoAdmin(boolean isNaoAdmin) {
+		this.isNaoAdmin = isNaoAdmin;
+	}
+
+	public boolean isAdmin() {
+		return isAdmin;
+	}
+
+	public void setAdmin(boolean isAdmin) {
+		this.isAdmin = isAdmin;
+	}
+
+	public Usuario getUsuarioAtualiza() {
+		return usuarioAtualiza;
+	}
+
+	public void setUsuarioAtualiza(Usuario usuarioAtualiza) {
+		this.usuarioAtualiza = usuarioAtualiza;
+	}
+
 }
